docs(feedback): remove unneeded file
