+++ packages/feedback/stories/feedback.stories.tsx
@@ -68,3 +68,61 @@ export const _Default: StoryObj<typeof Feedback> = {
     },
   },
 };
+
+export const _DefaultWithModal: StoryObj<typeof Feedback> = {
+  render: ({
+    btnText,
+    aboutOptions,
+    appName,
+    showSupport,
+    prompt,
+    additionalComments,
+    btnColor,
+    btnOutline,
+    className,
+    modal,
+  }) => (
+    <Feedback
+      appName={appName}
+      prompt={prompt}
+      formProps={{
+        additionalComments,
+        aboutOptions: aboutOptions
+          ? [
+              { label: 'The Payer Space', value: 'space' },
+              { label: 'Applications', value: 'applications' },
+              { label: 'Resources', value: 'resources' },
+              { label: 'News and Announcements', value: 'news' },
+            ]
+          : [],
+        staticFields: { staticKey: 'my-static-value' },
+      }}
+      showSupport={showSupport}
+      color={btnColor}
+      outline={btnOutline}
+      className={className}
+      modal={modal}
+    >
+      {btnText}
+    </Feedback>
+  ),
+  args: {
+    btnText: 'Button Text',
+    appName: 'Payer Space',
+    prompt: '',
+    additionalComments: false,
+    aboutOptions: false,
+    showSupport: false,
+    btnOutline: false,
+    btnColor: colorOptions[0],
+    className: '',
+    modal: true,
+  },
+  argTypes: {
+    btnColor: {
+      type: 'select',
+      options: colorOptions,
+    },
+  },
+  name: 'default w/ modal',
+};
